Connect frontend with backend

+++ frontend/src/template.rs
@@ -4,6 +4,6 @@ use askama::Template;
 #[template(path = "index.html")]
 pub struct Index<'a> {
     pub title: &'a str,
-    pub nodes: u8,
+    pub nodes: usize,
     pub kubernetes_version: &'a str,
 }
